Close #24: Add collection view for games lobby.
